# Implementing Snappy Block Encoding