--- a/pkg/util/grpcencoding/encoding_test.go
+++ b/pkg/util/grpcencoding/encoding_test.go
@@ -3,10 +3,12 @@ package grpcencoding
 import (
 	"bytes"
 	"io"
+	"io/ioutil"
 	"strings"
 	"testing"
 
 	"github.com/cortexproject/cortex/pkg/util/grpcencoding/snappy"
+	"github.com/cortexproject/cortex/pkg/util/grpcencoding/snappyblock"
 	"github.com/cortexproject/cortex/pkg/util/grpcencoding/zstd"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -19,6 +21,9 @@ func TestCompressors(t *testing.T) {
 	}{
 		{
 			name: snappy.Name,
+		},
+		{
+			name: snappyblock.Name,
 		},
 		{
 			name: zstd.Name,
@@ -42,30 +47,43 @@ func testCompress(name string, t *testing.T) {
 	}{
 		{"empty", ""},
 		{"short", "hello world"},
-		{"long", strings.Repeat("123456789", 1024)},
+		{"long", strings.Repeat("123456789", 2024)},
 	}
 	for _, test := range tests {
 		t.Run(test.test, func(t *testing.T) {
-			var buf bytes.Buffer
+			buf := &bytes.Buffer{}
 			// Compress
-			w, err := c.Compress(&buf)
+			w, err := c.Compress(buf)
 			require.NoError(t, err)
 			n, err := w.Write([]byte(test.input))
 			require.NoError(t, err)
 			assert.Len(t, test.input, n)
 			err = w.Close()
 			require.NoError(t, err)
+			compressedBytes := buf.Bytes()
+			buf = bytes.NewBuffer(compressedBytes)
+
 			// Decompress
-			r, err := c.Decompress(&buf)
+			r, err := c.Decompress(buf)
 			require.NoError(t, err)
 			out, err := io.ReadAll(r)
 			require.NoError(t, err)
 			assert.Equal(t, test.input, string(out))
+
+			if sizer, ok := c.(interface {
+				DecompressedSize(compressedBytes []byte) int
+			}); ok {
+				buf = bytes.NewBuffer(compressedBytes)
+				r, err := c.Decompress(buf)
+				require.NoError(t, err)
+				out, err := io.ReadAll(r)
+				assert.Equal(t, len(out), sizer.DecompressedSize(compressedBytes))
+			}
 		})
 	}
 }
 
-func BenchmarkSnappyCompress(b *testing.B) {
+func BenchmarkCompress(b *testing.B) {
 	data := []byte(strings.Repeat("123456789", 1024))
 
 	testCases := []struct {
@@ -73,6 +91,9 @@ func BenchmarkSnappyCompress(b *testing.B) {
 	}{
 		{
 			name: snappy.Name,
+		},
+		{
+			name: snappyblock.Name,
 		},
 		{
 			name: zstd.Name,
@@ -88,11 +109,12 @@ func BenchmarkSnappyCompress(b *testing.B) {
 				_, _ = w.Write(data)
 				_ = w.Close()
 			}
+			b.ReportAllocs()
 		})
 	}
 }
 
-func BenchmarkSnappyDecompress(b *testing.B) {
+func BenchmarkDecompress(b *testing.B) {
 	data := []byte(strings.Repeat("123456789", 1024))
 
 	testCases := []struct {
@@ -100,6 +122,9 @@ func BenchmarkSnappyDecompress(b *testing.B) {
 	}{
 		{
 			name: snappy.Name,
+		},
+		{
+			name: snappyblock.Name,
 		},
 		{
 			name: zstd.Name,
@@ -112,13 +137,40 @@ func BenchmarkSnappyDecompress(b *testing.B) {
 			var buf bytes.Buffer
 			w, _ := c.Compress(&buf)
 			_, _ = w.Write(data)
-			reader := bytes.NewReader(buf.Bytes())
+			w.Close()
 			b.ResetTimer()
 			for i := 0; i < b.N; i++ {
-				r, _ := c.Decompress(reader)
-				_, _ = io.ReadAll(r)
-				_, _ = reader.Seek(0, io.SeekStart)
+				_, _, err := decompress(c, buf.Bytes(), 10000)
+				require.NoError(b, err)
 			}
+			b.ReportAllocs()
 		})
 	}
+}
+
+// This function was copied from: https://github.com/grpc/grpc-go/blob/70c52915099a3b30848d0cb22e2f8951dd5aed7f/rpc_util.go#L765
+func decompress(compressor encoding.Compressor, d []byte, maxReceiveMessageSize int) ([]byte, int, error) {
+	dcReader, err := compressor.Decompress(bytes.NewReader(d))
+	if err != nil {
+		return nil, 0, err
+	}
+	if sizer, ok := compressor.(interface {
+		DecompressedSize(compressedBytes []byte) int
+	}); ok {
+		if size := sizer.DecompressedSize(d); size >= 0 {
+			if size > maxReceiveMessageSize {
+				return nil, size, nil
+			}
+			// size is used as an estimate to size the buffer, but we
+			// will read more data if available.
+			// +MinRead so ReadFrom will not reallocate if size is correct.
+			buf := bytes.NewBuffer(make([]byte, 0, size+bytes.MinRead))
+			bytesRead, err := buf.ReadFrom(io.LimitReader(dcReader, int64(maxReceiveMessageSize)+1))
+			return buf.Bytes(), int(bytesRead), err
+		}
+	}
+	// Read from LimitReader with limit max+1. So if the underlying
+	// reader is over limit, the result will be bigger than max.
+	d, err = ioutil.ReadAll(io.LimitReader(dcReader, int64(maxReceiveMessageSize)+1))
+	return d, len(d), err
 }