**What this PR does**:

This PR is a follow up of https://github.com/cortexproject/cortex/pull/5213#issuecomment-1474423754

This PR is implementing block format snappy encoding:

Benchmark Results:


```
pkg: github.com/cortexproject/cortex/pkg/util/grpcencoding
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkCompress/snappy-12         	  452925	      2339 ns/op	      16 B/op	       1 allocs/op
BenchmarkCompress/snappy-12         	  545708	      2264 ns/op	      16 B/op	       1 allocs/op
BenchmarkCompress/snappy-12         	  534312	      2230 ns/op	      16 B/op	       1 allocs/op
BenchmarkCompress/snappy-12         	  538390	      2293 ns/op	      16 B/op	       1 allocs/op
BenchmarkCompress/snappy-12         	  540451	      2201 ns/op	      16 B/op	       1 allocs/op
BenchmarkCompress/snappy-block-12   	  536930	      2233 ns/op	      16 B/op	       1 allocs/op
BenchmarkCompress/snappy-block-12   	  525919	      2306 ns/op	      16 B/op	       1 allocs/op
BenchmarkCompress/snappy-block-12   	  476133	      2271 ns/op	      16 B/op	       1 allocs/op
BenchmarkCompress/snappy-block-12   	  518916	      2203 ns/op	      16 B/op	       1 allocs/op
BenchmarkCompress/snappy-block-12   	  543210	      2187 ns/op	      16 B/op	       1 allocs/op
BenchmarkCompress/zstd-12           	  518970	      2202 ns/op	      16 B/op	       1 allocs/op
BenchmarkCompress/zstd-12           	  528393	      2227 ns/op	      16 B/op	       1 allocs/op
BenchmarkCompress/zstd-12           	  519684	      2455 ns/op	      16 B/op	       1 allocs/op
BenchmarkCompress/zstd-12           	  507280	      2268 ns/op	      16 B/op	       1 allocs/op
BenchmarkCompress/zstd-12           	  526838	      2234 ns/op	      16 B/op	       1 allocs/op
BenchmarkDecompress/snappy-12       	  115866	     10098 ns/op	   34212 B/op	      12 allocs/op
BenchmarkDecompress/snappy-12       	  117933	     10162 ns/op	   34254 B/op	      12 allocs/op
BenchmarkDecompress/snappy-12       	  107898	     10284 ns/op	   34238 B/op	      12 allocs/op
BenchmarkDecompress/snappy-12       	  116962	     10009 ns/op	   34262 B/op	      12 allocs/op
BenchmarkDecompress/snappy-12       	  115161	     10359 ns/op	   34237 B/op	      12 allocs/op
BenchmarkDecompress/snappy-block-12 	  192163	      6091 ns/op	    9858 B/op	       4 allocs/op
BenchmarkDecompress/snappy-block-12 	  192910	      6224 ns/op	    9859 B/op	       4 allocs/op
BenchmarkDecompress/snappy-block-12 	  191830	      6087 ns/op	    9857 B/op	       4 allocs/op
BenchmarkDecompress/snappy-block-12 	  193527	      6156 ns/op	    9859 B/op	       4 allocs/op
BenchmarkDecompress/snappy-block-12 	  191727	      6156 ns/op	    9858 B/op	       4 allocs/op
BenchmarkDecompress/zstd-12         	  112138	      9586 ns/op	   43905 B/op	      14 allocs/op
BenchmarkDecompress/zstd-12         	  119416	      9455 ns/op	   43903 B/op	      14 allocs/op
BenchmarkDecompress/zstd-12         	  121332	      9560 ns/op	   43903 B/op	      14 allocs/op
BenchmarkDecompress/zstd-12         	  118489	      9505 ns/op	   43903 B/op	      14 allocs/op
BenchmarkDecompress/zstd-12         	  117463	      9563 ns/op	   43902 B/op	      14 allocs/op
PASS
ok  	github.com/cortexproject/cortex/pkg/util/grpcencoding	36.997s
```

Comparing Snappy x snappy-block (Decompress)

```
name                  old time/op    new time/op    delta
Decompress/snappy-12    10.2µs ± 2%     6.1µs ± 1%  -39.67%  (p=0.008 n=5+5)

name                  old alloc/op   new alloc/op   delta
Decompress/snappy-12    34.2kB ± 0%     9.9kB ± 0%  -71.21%  (p=0.008 n=5+5)

name                  old allocs/op  new allocs/op  delta
Decompress/snappy-12      12.0 ± 0%       4.0 ± 0%  -66.67%  (p=0.008 n=5+5)
```

**Which issue(s) this PR fixes**:
Fixes #<issue number>

**Checklist**
- [x] Tests updated
- [x] Documentation added
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`