@alanprot 2023-03-18
The results seems pretty good. I dont see any change on the CPU and the network is 10X less.

![Screen Shot 2023-03-17 at 8 32 27 PM](https://user-images.githubusercontent.com/4027760/226082399-6c86af38-14cd-415d-8ec9-3be674b8d5d6.png)

![Screen Shot 2023-03-17 at 8 31 34 PM](https://user-images.githubusercontent.com/4027760/226082375-223e3034-eb8b-4501-bea8-2ddf2e9b923c.png)


The memory also seems better:
![Screen Shot 2023-03-17 at 8 34 30 PM](https://user-images.githubusercontent.com/4027760/226082570-8c01243f-ff3d-4a1c-bd30-11de4a9f8dcb.png)