@damnever 2023-03-22
@alanprot Yes, here is result of ingesters' CPU usage on our test environment (snappy/v1.14 vs snappy/#5213 vs snappy-block):
<img width="1291" alt="image" src="https://user-images.githubusercontent.com/6223594/226785963-d2a0616e-41c5-44b0-b548-0ed8943d5f83.png">