@alanprot 2023-03-22
Nice!! So if i'm reading correcly we have:

snappy/v1.14: 22.5% CPU
snappy/https://github.com/cortexproject/cortex/pull/5213 : 17.9% CPU 
snappy-block: 14.6% CPU

Seems that make sense to go ahead with this PR?

@yeya24 WDYT? 

Maybe snappy-block is not a good name haha :P