@damnever 2023-03-22
@alanprot You are right, but not in percentage rather CPU usage in seconds.