@alanprot 2023-03-22
Oh i see..

So it reduced the CPU usage in seconds by 45% (22.5*0.65=~14.6)? Nice!
We saw improvements on our test environments as well but not that big. This is good news!