@yeya24 2023-03-22
Nice! This is great improvement.
@alanprot About naming, I don't have strong preference. `snappy-block` also sounds good to me.